ml_service/requirements.txt
fastapi==0.116.1
uvicorn[standard]>=0.22,<1.0
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.1
loguru==0.7.3
pymupdf==1.26.4
pdfplumber==0.11.7
python-docx==1.2.0
scikit-learn==1.7.1
numpy==2.3.2
scipy==1.16.1
joblib==1.5.1
